Split authority import completion ownership
